Move CartoAPIError into carto-api-error.ts
